+++ frontend/lib/pages/inventory.dart
@@ -1,10 +1,10 @@
-import 'dart:math';
 
 import 'package:animated_text_kit/animated_text_kit.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:fl_chart/fl_chart.dart';
+import 'package:tooGoodToWaste/widgets/line_chart.dart';
 import 'dart:convert';
 import 'package:flutter/src/painting/gradient.dart' as Gradient; 
 
@@ -292,11 +292,15 @@ class _BottomTopScreenState extends State<Inventory>
   double expiringPercentage = 0.0;
   double goodPercentage = 0.0;
 
+  late bool isShowingMainData;
+
+
   @override
   void initState() {
     _tooltipBehavior = TooltipBehavior(enable: true);
  
     super.initState();
+    isShowingMainData = true;
     _tabController = TabController(length: 3, vsync: this);
     _fetchWasteNum();
     calculatePieChart();
@@ -369,7 +373,7 @@ class _BottomTopScreenState extends State<Inventory>
           return PieChartSectionData(
             color: Colors.red,
             value: wastedPercentage,
-            title: 'Wasted',
+            title: '${wastedPercentage.toStringAsFixed(2)}%',
             radius: radius,
             titleStyle: TextStyle(
               fontSize: fontSize,
@@ -382,7 +386,7 @@ class _BottomTopScreenState extends State<Inventory>
           return PieChartSectionData(
             color: Colors.blue,
             value: goodPercentage,
-            title: 'Used',
+            title: '${goodPercentage.toStringAsFixed(2)}%',
             radius: radius,
             titleStyle: TextStyle(
               fontSize: fontSize,
@@ -395,7 +399,7 @@ class _BottomTopScreenState extends State<Inventory>
           return PieChartSectionData(
             color: Colors.green,
             value: sharedPercentage,
-            title: 'Shared',
+            title: '${sharedPercentage.toStringAsFixed(2)}%',
             radius: radius,
             titleStyle: TextStyle(
               fontSize: fontSize,
@@ -408,7 +412,7 @@ class _BottomTopScreenState extends State<Inventory>
           return PieChartSectionData(
             color: Colors.yellow,
             value: expiringPercentage,
-            title: 'Expiring',
+            title: '${expiringPercentage.toStringAsFixed(2)}%',
             radius: radius,
             titleStyle: TextStyle(
               fontSize: fontSize,
@@ -422,6 +426,13 @@ class _BottomTopScreenState extends State<Inventory>
       }
     });
   }
+
+  List<String> monthNames = [
+    '', // Index 0 is empty since months start from 1
+    'January', 'February', 'March', 'April', 'May', 'June', 'July', 'August',
+    'September', 'October', 'November', 'December'
+  ];
+
     
     return Scaffold(
       appBar: AppBar(
@@ -481,9 +492,9 @@ class _BottomTopScreenState extends State<Inventory>
                         const SizedBox(
                           height: 5,
                         ),
-                        Text('Month: ${timeNowDate.month} ${timeNowDate.year}'),
+                        Text('Month: ${monthNames[timeNowDate.month]} ${timeNowDate.year}'),
                         const SizedBox(
-                          height: 10,
+                          height: 20,
                         ),
                     Row(
                       children: <Widget>[
@@ -579,17 +590,29 @@ class _BottomTopScreenState extends State<Inventory>
                         ),
                         const Text('Period: April 2024 - May 2024'),
                         const SizedBox(
-                          height: 10,
+                          height: 20,
                         ),
+                        IconButton(
+                                  icon: Icon(
+                                    Icons.refresh,
+                                    color: Colors.white.withOpacity(isShowingMainData ? 1.0 : 0.5),
+                                  ),
+                                  onPressed: () {
+                                    setState(() {
+                                      isShowingMainData = !isShowingMainData;
+                                    });
+                                  },
+                                  ),
                         SizedBox(
                             height: 200,
                             child: Row(
                               mainAxisAlignment: MainAxisAlignment.center,
                               children: [
+                               
                                 SizedBox(
                                   width: MediaQuery.of(context).size.width -
                                       20, // 20 is padding left + right
-                                  child: const _LineChart(),
+                                  child: Line_Chart(isShowingMainData: isShowingMainData),
                                 )
                               ],
                             )),
@@ -1475,235 +1498,3 @@ class _BottomTopScreenState extends State<Inventory>
   }
 }
 
-class _LineChart extends StatelessWidget {
-  const _LineChart();
-
-  @override
-  Widget build(BuildContext context) {
-    return LineChart(
-      sampleData2,
-      duration: const Duration(milliseconds: 250),
-    );
-  }
-
-  LineChartData get sampleData2 => LineChartData(
-        lineTouchData: lineTouchData2,
-        gridData: gridData,
-        titlesData: titlesData2,
-        lineBarsData: lineBarsData2,
-        minX: 0,
-        maxX: 13,
-        maxY: 17,
-        minY: 0,
-      );
-
-  LineTouchData get lineTouchData1 => LineTouchData(
-        handleBuiltInTouches: true,
-        touchTooltipData: LineTouchTooltipData(
-
-          //tooltipBgColor: Colors.blueGrey.withOpacity(0.8),
-        ),
-      );
-
-  FlTitlesData get titlesData1 => FlTitlesData(
-        bottomTitles: AxisTitles(
-          sideTitles: bottomTitles,
-        ),
-        rightTitles: const AxisTitles(
-          sideTitles: SideTitles(showTitles: false),
-        ),
-        topTitles: const AxisTitles(
-          sideTitles: SideTitles(showTitles: false),
-        ),
-        leftTitles: AxisTitles(
-          sideTitles: leftTitles(),
-        ),
-      );
-
-  LineTouchData get lineTouchData2 => const LineTouchData(
-        enabled: false,
-      );
-
-  FlTitlesData get titlesData2 => FlTitlesData(
-        bottomTitles: AxisTitles(
-          sideTitles: bottomTitles,
-        ),
-        rightTitles: const AxisTitles(
-          sideTitles: SideTitles(showTitles: false),
-        ),
-        topTitles: const AxisTitles(
-          sideTitles: SideTitles(showTitles: false),
-        ),
-        leftTitles: AxisTitles(
-          sideTitles: leftTitles(),
-        ),
-      );
-
-  List<LineChartBarData> get lineBarsData2 => [
-        lineChartBarData2_1,
-        lineChartBarData2_2,
-        lineChartBarData2_3,
-      ];
-
-  Widget leftTitleWidgets(double value, TitleMeta meta) {
-    const style = TextStyle(
-      fontSize: 10,
-    );
-    String text;
-    switch (value.toInt()) {
-      case 5:
-        text = '5kg';
-        break;
-      case 10:
-        text = '10kg';
-        break;
-      case 15:
-        text = '15kg';
-        break;
-      default:
-        return Container();
-    }
-
-    return Text(text, style: style, textAlign: TextAlign.center);
-  }
-
-  SideTitles leftTitles() => SideTitles(
-        getTitlesWidget: leftTitleWidgets,
-        showTitles: true,
-        interval: 1,
-        reservedSize: 40,
-      );
-
-  Widget bottomTitleWidgets(double value, TitleMeta meta) {
-    const style = TextStyle(
-      fontSize: 8,
-    );
-    Widget text;
-    switch (value.toInt()) {
-      case 1:
-        text = const Text('31', style: style);
-        break;
-      case 2:
-        text = const Text('2', style: style);
-        break;
-      case 3:
-        text = const Text('5', style: style);
-        break;
-      case 4:
-        text = const Text('8', style: style);
-        break;
-      case 5:
-        text = const Text('11', style: style);
-        break;
-      case 6:
-        text = const Text('14', style: style);
-        break;
-      case 7:
-        text = const Text('17', style: style);
-        break;
-      case 8:
-        text = const Text('20', style: style);
-        break;
-      case 9:
-        text = const Text('23', style: style);
-        break;
-      case 10:
-        text = const Text('26', style: style);
-        break;
-      case 11:
-        text = const Text('29', style: style);
-        break;
-      case 12:
-        text = const Text('3', style: style);
-        break;
-      default:
-        text = const Text('');
-        break;
-    }
-
-    return SideTitleWidget(
-      axisSide: meta.axisSide,
-      space: 10,
-      child: text,
-    );
-  }
-
-  SideTitles get bottomTitles => SideTitles(
-        showTitles: true,
-        reservedSize: 36,
-        interval: 1,
-        getTitlesWidget: bottomTitleWidgets,
-      );
-
-  FlGridData get gridData => const FlGridData(show: false);
-
-  LineChartBarData get lineChartBarData2_1 => LineChartBarData(
-        isCurved: true,
-        curveSmoothness: 0,
-        color: Colors.blue,
-        barWidth: 1,
-        isStrokeCapRound: true,
-        dotData: const FlDotData(show: false),
-        belowBarData: BarAreaData(show: false),
-        spots: const [
-          FlSpot(1, 1),
-          FlSpot(2, 3),
-          FlSpot(3, 4),
-          FlSpot(4, 2),
-          FlSpot(5, 1.8),
-          FlSpot(6, 4),
-          FlSpot(7, 7),
-          FlSpot(8, 5),
-          FlSpot(9, 3),
-          FlSpot(10, 2),
-          FlSpot(11, 1),
-          FlSpot(12, 2.2),
-        ],
-      );
-
-  LineChartBarData get lineChartBarData2_2 => LineChartBarData(
-        isCurved: true,
-        curveSmoothness: 0,
-        color: Colors.green,
-        barWidth: 1,
-        isStrokeCapRound: true,
-        dotData: const FlDotData(show: false),
-        spots: const [
-          FlSpot(1, 12),
-          FlSpot(2, 11),
-          FlSpot(3, 14),
-          FlSpot(4, 9),
-          FlSpot(5, 11),
-          FlSpot(6, 10),
-          FlSpot(7, 12),
-          FlSpot(8, 13),
-          FlSpot(9, 14),
-          FlSpot(10, 11),
-          FlSpot(11, 10),
-          FlSpot(12, 12),
-        ],
-      );
-
-  LineChartBarData get lineChartBarData2_3 => LineChartBarData(
-        isCurved: true,
-        curveSmoothness: 0,
-        color: Colors.red,
-        barWidth: 1,
-        isStrokeCapRound: true,
-        dotData: const FlDotData(show: false),
-        spots: const [
-          FlSpot(1, 3.8),
-          FlSpot(2, 2),
-          FlSpot(3, 1.9),
-          FlSpot(4, 2),
-          FlSpot(5, 1),
-          FlSpot(6, 5),
-          FlSpot(7, 1),
-          FlSpot(8, 2),
-          FlSpot(9, 4),
-          FlSpot(10, 3.3),
-          FlSpot(11, 2),
-          FlSpot(12, 4),
-        ],
-      );
-}
